Remove contrib extensions from Imply Druid (#40)
* Remove contrib extensions from Imply Druid

Imply does not support any contrib extensions. Remove all of them so that
the build is faster and we don't need to deal with security vulnerabilities
flagged in extensions that are not packaged with Imply

* Revert "Remove contrib extensions from Imply Druid"

This reverts commit f22e0c275e6a43d585ef3b49e41f4e2ba84c88c8.

* Remove contrib extensions from Imply Druid

Imply does not support any contrib extensions. Remove all of them so that
the build is faster and we don't need to deal with security vulnerabilities
flagged in extensions that are not packaged with Imply

The code for the contrib extensions are left in the repo so we do not have
to resolve any conflicts that arise from new changes to the contrib extensions

* fix unused declaration warnings

* oops