Auto sync monthly (#593)
* Auto sync monthly

Get rid of the PR step - there have been limited commits that have needed
to be reverted over the last year of auto-backports. Now, the release
manager is responsible for reverting backward incompatible changes from the
monthly branch.

There will be 2 active branches - `master` for continuous updates from the open
source and `monthly` for all updates minus incompatible changes accepted in the
open source

eg. ZK 3.5 support in the middle of the year that is introduced in a
backwards incompatible change. This commit will get sync'd with master, then
sync'd with monthly. The tests / release manager detect the incompatible change
and revert it from the monthly branch. When the time comes to cut the next LTS
branch, it will be cut from master where the backwards incompatible change
exists, hence picking it up only on the next LTS release.

* Apply suggestions from code review