[Imply-26106] Updates Jenkins pipeline to run cloud deep-storage tests and license checks with Hadoop3 build (#1488)
Hadoop 2 often causes red security scans on imply bundles because of the dependencies it brings. We want to move away from Hadoop 2 and have to provide Hadoop 3 distribution available.

New stages are added in the pipeline to make the license checks pipeline, and existing IT tests that run with Hadoop2 also run with the Hadoop3 build.