Fix broken external-schema extension tests
Fixed tests in the external-schema extension that were broken by a recent
apache change to the input source security feature. The feature is experimental
and undocumented. The breakage was caused by the switching of the name, and type
parameters of the resource action needed for the input sourc. Updated implementation
code and tests as needed.

Also fixed an unrelated test compilation error in the s3 extension