IMPLY-29829: Keycloak authenticator for API keys (#1751)
Fixes IMPLY-29829.
Adds a new authenticator for Imply API keys to the imply-keycloak extension.

The authenticator currently only implements JDBC authentication. It has a dummy pass through filter so it can't be used for other API paths. Under the hood, the authenticator uses a helper class that does some simple LRU caching of TokenManagers so we get the just-in-time token refresh stuff for free

Key changed/added classes in this PR
* ImplyAPIKeyAuthenticator
* ImplyKeycloakAPIKeyAuthHelper
* ImplyKeycloakModule