Integrate with jfr-profiler
Add dependency on io.imply:jfr_profiler library. This optionally starts
a continuous, jfr streaming-based profiler with any clarity http
emitter.

Adds system property 'druid.emitter.clarity.jfrProfiler' which defaults
to false. Enabling it enables the jfr-profiler. The profiler only works
when run in a jvm17+, so even with the option on, it will not load on
older jvms.