StringMatch: Set limit before position on buffers. (#2193)
Important because the grouping engine may, in some cases, pass in
a buffer whose position is beyond the position of the current aggregation.
In this case, when we restore the position to the originalPosition in
the finally block, the originalPosition may be well past the shortened
limit (stringPosition + maxLength). To avoid this, restore the limit
first, then the position.